Update domain model and use new orc interfaces
We've discovered that the NANDO database doesn't
model procedures as first class entities. Only the
annex that contains a procedure has a unique ID.
For this reason, we've changed the table structure
to represent annexes rather than procedures.

An additional complexity is that annex IDs are only
unique based on their name, regardless of the
legislation to which they apply, so we have used a
synthesised compound key based on the legislation
and annex ID to make each annex tied to a specific
legislation, and therefore able to hold procedures.

We have also changed the interface with `orc` to
use a prototype editing interface, which allows
update of an existing store rather fresh creation
of RSF each time. This is more suitable for update
of existing stores. It also means RSF gets emitted
with the correct hash assertions.